requirements.txt
pillow==10.3.0
flake8>=3.9.0,<4.0.1
black==22.3.0
pytest>=7.0.0,<=7.4.2
drf-yasg==1.21.4
phonenumbers==8.13.35
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
hijri-converter==2.3.1

# django
Django==5.0
djangorestframework==3.15.1
django-geojson==4.0.0
django-filter==24.2
pytest-django==4.8
djangorestframework-simplejwt==5.3.1
